[UPDATED] refresh token

+++ src/modules/auth/auth.service.js
@@ -1,59 +1,28 @@
-import httpStatus from 'http-status';
 import bcrypt from 'bcryptjs';
+import httpStatus from 'http-status';
+
+import sendServiceResponse from '../../utils/sendServiceResponse.js';
+import newServiceErrorHandler from '../../utils/newServiceErrorHandler.js';
 
 import tokenService from './token/token.service.js';
 import userService from '../user/user.service.js';
 import TokenService from './token/token.service.js';
 import TokenModel from './token/token.model.js';
-import { tokenTypes } from '../../config/tokens.config.js';
+import tokenTypes from '../../config/tokens.config.js';
+import RoleAggregationPipeline from './role/role.pipeline.js';
+import RoleModel from './role/role.model.js';
+import UserModel from '../user/user.model.js';
 
-import ServerError from '../../utils/serverError.js';
-
-/**
- * Login with username and password
- * @param {string} email
- * @param {string} password
- * @returns {Promise<User>}
- */
 const loginUserWithEmailAndPassword = async (email, password) => {
     try {
         const userDetails = await userService.getUserByEmail(email);
 
-        if (!userDetails || userDetails.email !== email) {
-            return {
-                serviceSuccess: false,
-                serviceStatus: httpStatus.UNAUTHORIZED,
-                serviceMessage: 'Wrong email or password',
-                serviceData: {},
+        if (!userDetails || userDetails?.email !== email) {
+            throw {
+                statusCode: httpStatus.UNAUTHORIZED,
+                message: 'Wrong email or password',
             };
         }
-
-        // if (userDetails.failedLoginDetails?.length >= MAX_FAILED_LOGIN_ATTEMPTS) {
-        //     return {
-        //         serviceSuccess: false,
-        //         serviceStatus: httpStatus.LOCKED,
-        //         serviceMessage: 'Your account has been locked. Please contact support.',
-        //         serviceData: {},
-        //     };
-        // }
-        //
-        // if (userDetails?.mustChangePassword) {
-        //     return {
-        //         serviceSuccess: false,
-        //         serviceStatus: httpStatus.FORBIDDEN,
-        //         serviceMessage: 'Please change your password before proceeding.',
-        //         serviceData: {},
-        //     };
-        // }
-        //
-        // if (userDetails.activeSessionDetails?.length >= MAX_CONCURRENT_LOGINS) {
-        //     return {
-        //         serviceSuccess: false,
-        //         serviceStatus: httpStatus.FORBIDDEN,
-        //         serviceMessage: `Maximum concurrent login device limit reached. You can log in a maximum ${MAX_CONCURRENT_LOGINS} device at a time. Please log out from another device before logging in.`,
-        //         serviceData: {},
-        //     };
-        // }
 
         const passwordMatch = await bcrypt.compare(
             password,
@@ -61,55 +30,65 @@ const loginUserWithEmailAndPassword = async (email, password) => {
         );
 
         if (!passwordMatch) {
-            // await updateUserWithFailedLoginDetails(userDetails, device);
-
-            return {
-                serviceSuccess: false,
-                serviceStatus: httpStatus.UNAUTHORIZED,
-                serviceMessage: 'Wrong email or password',
-                serviceData: {},
+            throw {
+                statusCode: httpStatus.UNAUTHORIZED,
+                message: 'Wrong email or password',
             };
         }
 
-        const { serviceData } =
-            await TokenService.generateAuthTokens(userDetails);
+        const token = await TokenService.generateAuthTokens(userDetails);
 
         // Convert the Mongoose document to a plain JavaScript object
         let userData = userDetails.toObject();
 
-        // Remove the password field from the object
-        delete userData.password;
+        // Aggregation pipeline to fetch and populate the updated document
+        const aggregationPipeline = RoleAggregationPipeline.getRole(
+            userDetails?.role
+        );
 
-        return {
-            serviceSuccess: true,
-            serviceStatus: serviceData
-                ? httpStatus.OK
-                : httpStatus.UNAUTHORIZED,
-            serviceMessage: serviceData
-                ? 'Login successful.'
-                : 'Invalid email or password',
-            serviceData: {
+        // Fetch the updated permission using the aggregation pipeline
+        const populatedPermission =
+            await RoleModel.aggregate(aggregationPipeline);
+
+        // Check if the populatedPermission query returned a document
+        if (!populatedPermission || populatedPermission?.length === 0) {
+            throw {
+                statusCode: httpStatus.OK,
+                message: 'User login successful but role population failed.',
+            };
+        }
+
+        // Remove the password field from the object
+        delete userData._id;
+        delete userData.__v;
+        delete userData.password;
+        delete userData.role;
+
+        return sendServiceResponse(
+            token ? httpStatus.OK : httpStatus.UNAUTHORIZED,
+            token ? 'Login successful.' : 'Invalid email or password',
+            {
                 ...userData,
-                token: serviceData,
-            },
-        };
+                role: populatedPermission[0],
+                token,
+            }
+        );
     } catch (error) {
-        return {
-            serviceSuccess: false,
-            serviceData: {},
-            serviceMessage: 'Internal server error',
-            serviceStatus: httpStatus.INTERNAL_SERVER_ERROR,
-        };
+        // Handle errors that might occur during the user creation process
+        return newServiceErrorHandler(error);
     }
 };
 
-/**
- * Logout
- * @param {string} refreshToken
- * @returns {Promise}
- */
 const logout = async refreshToken => {
     try {
+        // Verify if refreshToken is provided
+        if (!refreshToken) {
+            throw {
+                statusCode: httpStatus.BAD_REQUEST,
+                message: 'Refresh token is required.',
+            };
+        }
+
         const refreshTokenDoc = await TokenModel.findOneAndDelete({
             token: refreshToken,
             type: tokenTypes.REFRESH,
@@ -118,113 +97,80 @@ const logout = async refreshToken => {
 
         // If no document is found, throw an error to indicate the token was not found
         if (!refreshTokenDoc) {
-            throw new ServerError(httpStatus.NOT_FOUND, 'Not found');
-        }
-
-        // Return a success response after successfully removing the token
-        return {
-            serviceSuccess: true,
-            serviceStatus: httpStatus.OK,
-            serviceMessage: 'Logout successful.',
-            serviceData: {}, // refreshTokenDoc is removed, so we might not want to return it
-        };
-    } catch (error) {
-        return {
-            serviceSuccess: false,
-            serviceData: {},
-            serviceMessage: error.message || 'Internal server error', // Providing dynamic error messages based on the caught error
-            serviceStatus: error.status || httpStatus.INTERNAL_SERVER_ERROR, // Using the error's status if available, otherwise default to 500
-        };
-    }
-};
-
-/**
- * Refresh auth tokens
- * @param {string} refreshToken
- * @returns {Promise<Object>}
- */
-const refreshAuth = async refreshToken => {
-    try {
-        // Verify the token and get its associated data
-        const verificationResult = await tokenService.verifyToken(
-            refreshToken,
-            tokenTypes.REFRESH
-        );
-
-        // Fetch the user details associated with the token
-        const userDetails = await userService.getUserById(
-            verificationResult.serviceData.user
-        );
-
-        if (!userDetails) {
-            return {
-                serviceSuccess: false,
-                serviceStatus: httpStatus.UNAUTHORIZED,
-                serviceMessage: 'User not found with the refresh token.',
-                serviceData: {},
+            throw {
+                statusCode: httpStatus.NOT_FOUND,
+                message: 'Token not found or already logged out.',
             };
         }
 
-        // Fetch the token document by its ID
-        const tokenDoc = await TokenModel.findOneAndDelete({
-            user: verificationResult.serviceData.user,
-        });
-
-        if (!tokenDoc) {
-            throw new Error('Token not found or already removed.');
-        }
-
-        // Generate new auth tokens
-        const authTokensResult =
-            await tokenService.generateAuthTokens(userDetails);
-
-        return {
-            serviceSuccess: true,
-            serviceStatus: httpStatus.OK,
-            serviceMessage: 'Refresh successful.',
-            serviceData: authTokensResult.serviceData, // Correctly extracting serviceData from the result
-        };
+        return sendServiceResponse(httpStatus.OK, 'Logout successful.', {});
     } catch (error) {
-        return {
-            serviceSuccess: false,
-            serviceData: {},
-            serviceMessage: 'Internal server error',
-            serviceStatus: httpStatus.INTERNAL_SERVER_ERROR,
-        };
+        // Handle errors that might occur during the user creation process
+        return newServiceErrorHandler(error);
     }
 };
 
-/**
- * Reset password
- * @param {string} resetPasswordToken
- * @param {string} newPassword
- * @returns {Promise}
- */
+const refreshAuth = async refreshToken => {
+    try {
+        const refreshTokenDoc = await TokenModel.findOne({
+            token: refreshToken,
+            type: tokenTypes.REFRESH,
+        });
+
+        const user = await UserModel.findOne({ id: refreshTokenDoc.user });
+
+        if (!user) {
+            throw new Error();
+        }
+
+        await TokenModel.findOneAndDelete({ token: refreshToken });
+
+        const newAuthToken = await tokenService.generateAuthTokens(user);
+
+        return sendServiceResponse(
+            httpStatus.OK,
+            'Refresh successful.',
+            newAuthToken
+        );
+    } catch (error) {
+        // Handle errors that might occur during the user creation process
+        return newServiceErrorHandler(error);
+    }
+};
+
 const resetPassword = async (resetPasswordToken, newPassword) => {
     try {
         const resetPasswordTokenDoc = await tokenService.verifyToken(
             resetPasswordToken,
             tokenTypes.RESET_PASSWORD
         );
+
         const user = await userService.getUserById(resetPasswordTokenDoc.user);
+
         if (!user) {
-            throw new Error();
+            throw {
+                statusCode: httpStatus.NOT_FOUND,
+                message: 'User not found with the reset password token.',
+            };
         }
+
         await userService.updateUserById(user.id, { password: newPassword });
         await TokenModel.deleteMany({
             user: user.id,
             type: tokenTypes.RESET_PASSWORD,
         });
+
+        return sendServiceResponse(
+            httpStatus.OK,
+            'Password reset successful.',
+            {}
+        );
     } catch (error) {
-        throw new ServerError(httpStatus.UNAUTHORIZED, 'Password reset failed');
+        // Handle errors that might occur during the user creation process
+        return newServiceErrorHandler(error);
     }
 };
 
-/**
- * Verify email
- * @param {string} verifyEmailToken
- * @returns {Promise}
- */
 const verifyEmail = async verifyEmailToken => {
     try {
         const verifyEmailTokenDoc = await tokenService.verifyToken(
@@ -236,31 +182,29 @@ const verifyEmail = async verifyEmailToken => {
         );
 
         if (!userDetails) {
-            return {
-                serviceSuccess: false,
-                serviceData: {},
-                serviceMessage: 'User not found with the verify email token.',
-                serviceStatus: httpStatus.FORBIDDEN,
+            throw {
+                statusCode: httpStatus.FORBIDDEN,
+                message: 'User not found with the verify email token.',
             };
         }
 
-        console.log(userDetails);
-
         await TokenModel.deleteMany({
-            user: userDetails._id,
+            user: userDetails.id,
             type: tokenTypes.VERIFY_EMAIL,
         });
 
         await userService.updateUserById(userDetails.id, {
             isEmailVerified: true,
         });
+
+        return sendServiceResponse(
+            httpStatus.OK,
+            'Email verification successful.',
+            {}
+        );
     } catch (error) {
-        return {
-            serviceSuccess: false,
-            serviceData: {},
-            serviceMessage: 'Email verification failed',
-            serviceStatus: httpStatus.UNAUTHORIZED,
-        };
+        // Handle errors that might occur during the user creation process
+        return newServiceErrorHandler(error);
     }
 };
